requirements.txt
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
fastapi-sso==0.15.0
filelock==3.16.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.35.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
midtransclient==1.4.2
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
pyasn1==0.6.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.3
tifffile==2024.9.20
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.24.0
websockets==13.1
XlsxWriter==3.2.0
pillow==10.4.0
pytz==2024.2
reportlab==4.2.5